// src/app/api-cinekeep/model/episodeRatingRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SeriesSnapshotRequest } from './seriesSnapshotRequest';


export interface EpisodeRatingRequest { 
    value: number;
    episodeName: string;
    stillPath?: string;
    airDate?: string;
    series: SeriesSnapshotRequest;
}