// src/app/api-cinekeep/model/seriesSnapshotRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SeriesSnapshotRequest { 
    title: string;
    posterPath?: string;
    backdropPath?: string;
    overview?: string;
    releaseDate?: string;
    voteAverage?: number;
    voteCount?: number;
}