// src/app/api-cinekeep/model/watchlistItemResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WatchlistItemResponse { 
    id?: number;
    tmdbId?: number;
    mediaType?: WatchlistItemResponse.MediaTypeEnum;
    title?: string;
    posterPath?: string;
    backdropPath?: string;
    overview?: string;
    releaseDate?: string;
    voteAverage?: number;
    voteCount?: number;
    addedAt?: string;
}
export namespace WatchlistItemResponse {
    export const MediaTypeEnum = {
        Movie: 'movie',
        Tv: 'tv'
    } as const;
    export type MediaTypeEnum = typeof MediaTypeEnum[keyof typeof MediaTypeEnum];
}